% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aqc}
\alias{aqc}
\title{Aquilegia search results}
\format{
An character string with 107 articles.
}
\source{
NCBI PubMed
}
\usage{
aqc
}
\description{
Aquilegia search results
}
\examples{
\dontrun{
res <- pubmed_search("aquilegia[TITLE]")
aqc <- pubmed_fetch(res, parsed=FALSE)
}
str(aqc)
aq <- pubmed_nodeset(aqc)
aq
cat(as.character(aq[1]))
}
\keyword{datasets}
