% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dbc}
\alias{dbc}
\title{PubMed articles with multiple databanks and accession numbers}
\format{
An character string with 3 articles.
}
\source{
NCBI PubMed
}
\usage{
dbc
}
\description{
PubMed articles with multiple databanks and accession numbers
}
\examples{
\dontrun{
res <- pubmed_search("31566309[PMID] OR 16939956[PMID] OR 30145791[PMID]")
dbc <- pubmed_fetch(res, parsed=FALSE)
}
str(dbc)
db <- pubmed_nodeset(dbc)
db
cat(as.character(db[1]))
}
\keyword{datasets}
