% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pubmed.R
\name{extract_pubmed}
\alias{extract_pubmed}
\title{Extract datatypes from PubMed XML records}
\usage{
extract_pubmed(
  filepath,
  datatype = NULL,
  pmid = NULL,
  processed_dir = NULL,
  quiet = FALSE
)
}
\arguments{
\item{filepath}{Character. Filepath including directory portion. Filepath should end in ".xml". If \code{pmid} is not provided, filename should be PMID.}

\item{datatype}{Type of data to extract from xml for which there is a corresponding "pubmed_" function ("main", "abstract", "databanks", "authors", "mesh", "keywords", "pubtypes"). "main" used `pubmed_table()`}

\item{pmid}{Character. Default = NULL in which case \code{pmid} parsed from the filename.}

\item{processed_dir}{Character. Default = NULL. Filepath of a directory to which processed file should be moved. By default, file will not be moved.}

\item{quiet}{Logical. Default = FALSE. Whether to inform user.}
}
\value{
Dataframe with output from corresponding "pubmed_" function
}
\description{
Extract datatypes from PubMed XML records
}
