% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_abstract.R
\name{pubmed_abstract}
\alias{pubmed_abstract}
\title{Parse abstract}
\usage{
pubmed_abstract(nodes, sentence = FALSE)
}
\arguments{
\item{nodes}{article node set}

\item{sentence}{split into sentences}
}
\value{
A tibble
}
\description{
Parse abstract
}
\examples{
# can't save xml_nodesets in R/data
aq <- pubmed_nodeset(aqc)
x <- pubmed_abstract(aq)
x
mutate(x, text=ifelse(is.na(label), abstract, paste0(label, ": ", abstract))) \%>\%
  group_by(pmid) \%>\%
  summarize(abstract=paste(text, collapse=" ")) \%>\%
  arrange(desc(pmid))
pubmed_abstract(aq, sentence=TRUE)
}
