% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_authors.R
\name{pubmed_authors}
\alias{pubmed_authors}
\title{Parse authors}
\usage{
pubmed_authors(nodes)
}
\arguments{
\item{nodes}{article node set}
}
\value{
A tibble
}
\description{
Parse authors
}
\examples{
aq <- pubmed_nodeset(aqc)
x <- pubmed_authors(aq)
x
mutate(x, name=paste(last, initials)) \%>\%
  group_by(pmid) \%>\%
  summarize(authors=paste(name, collapse=", "))
mutate(x, name=ifelse(lead(n) == 5, "et al", paste(last, initials))) \%>\%
  filter(n < 5) \%>\%
  group_by(pmid) \%>\%
  summarize(authors=paste(name, collapse=", "))
}
