% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_databanks.R
\name{pubmed_databanks}
\alias{pubmed_databanks}
\title{Parse databanks}
\usage{
pubmed_databanks(nodes)
}
\arguments{
\item{nodes}{article node set}
}
\value{
A tibble
}
\description{
Parse databanks
}
\examples{
aq <- pubmed_nodeset(aqc)
x <- pubmed_databanks(aq)
x
count(x, pmid, databank, sort=TRUE)

db <- pubmed_nodeset(dbc)
y <- pubmed_databanks(db)
y
count(y, pmid, databank, sort=TRUE)
}
