% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_fetch.R
\name{pubmed_fetch}
\alias{pubmed_fetch}
\title{Fetch PubMed results}
\usage{
pubmed_fetch(history, parsed = TRUE)
}
\arguments{
\item{history}{web history results from \code{\link{pubmed_search}}}

\item{parsed}{parse results into article xml_nodeset with PMID names, default TRUE}
}
\value{
A character string or xml_nodeset if parsed=TRUE
}
\description{
Run \code{entrez_fetch} and return article nodes
}
\examples{
res <- pubmed_search("aquilegia[TITLE]")
aq <- pubmed_fetch(res)
aq
cat(as.character(aq[1]))
}
