% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_keywords.R
\name{pubmed_keywords}
\alias{pubmed_keywords}
\title{Parse keywords}
\usage{
pubmed_keywords(nodes)
}
\arguments{
\item{nodes}{article node set}
}
\value{
A tibble
}
\description{
Parse keywords
}
\examples{
aq <- pubmed_nodeset(aqc)
x <- pubmed_keywords(aq)
x
count(x, keyword, sort=TRUE)
}
