% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_mesh.R
\name{pubmed_mesh}
\alias{pubmed_mesh}
\title{Parse MeSH terms}
\usage{
pubmed_mesh(nodes)
}
\arguments{
\item{nodes}{article node set}
}
\value{
A tibble
}
\description{
Parse MeSH terms
}
\examples{
aq <- pubmed_nodeset(aqc)
x <- pubmed_mesh(aq)
x
mutate(x, mesh=gsub("\\\\*", "", mesh)) \%>\%
  count(mesh, sort=TRUE)
unique(x[, 1:3]) \%>\%
  count(descriptor, sort=TRUE)
}
