% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_nodeset.R
\name{pubmed_nodeset}
\alias{pubmed_nodeset}
\title{Parse PubMed into xml_nodeset with PMID names}
\usage{
pubmed_nodeset(esearch)
}
\arguments{
\item{esearch}{a character string from \code{link{pubmed_search}}}
}
\value{
A xml_nodeset
}
\description{
Parse PubMed into xml_nodeset with PMID names
}
\examples{
\dontrun{
res <- pubmed_search("aquilegia[TITLE]")
aqc <- pubmed_fetch(res, parsed=FALSE)
}
aq <- pubmed_nodeset(aqc)
aq
cat(as.character(aq[1]))
}
