% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_pubtypes.R
\name{pubmed_pubtypes}
\alias{pubmed_pubtypes}
\title{Parse publication types}
\usage{
pubmed_pubtypes(nodes)
}
\arguments{
\item{nodes}{article node set}
}
\value{
A tibble
}
\description{
Parse publication types
}
\examples{
aq <- pubmed_nodeset(aqc)
x <- pubmed_pubtypes(aq)
x
count(x, publication_type, sort=TRUE)
}
