% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{pubmed_search}
\alias{pubmed_search}
\title{Search PubMed}
\usage{
pubmed_search(term)
}
\arguments{
\item{term}{search term}
}
\value{
A list
}
\description{
Run \code{entrez_search} and return the web history.
}
\examples{
res <- pubmed_search("aquilegia[TITLE]")
res
# aq <- pubmed_fetch(res)
}
