% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_table.R
\name{pubmed_table}
\alias{pubmed_table}
\title{Parse article metadata}
\usage{
pubmed_table(nodes, etal = FALSE, iso = FALSE)
}
\arguments{
\item{nodes}{article node set}

\item{etal}{shorten author string, default false}

\item{iso}{use journal ISO abbreviation, default false}
}
\value{
A tibble
}
\description{
Parse article metadata
}
\examples{
aq <- pubmed_nodeset(aqc)
x <- pubmed_table(aq)
x
}
