% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypubmed-package.R
\docType{package}
\name{tidypubmed}
\alias{tidypubmed}
\title{\code{tidypubmed} package}
\description{
Parse XML documents from PubMed
}
\details{
See the Github page for details at \url{https://github.com/ropensci/tidypubmed}
}
\keyword{internal}
