% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_text_first.R
\name{xml_text_first}
\alias{xml_text_first}
\alias{xml_attr_first}
\title{Extract first xml text}
\usage{
xml_text_first(nodes, xpath)

xml_attr_first(nodes, xpath, attr)
}
\arguments{
\item{nodes}{article node set with PMID names}

\item{xpath}{xml node name}

\item{attr}{xml attribute name}
}
\value{
A vector. Used by \code{pubmed_table} to add additional columns
}
\description{
Extract first xml text
}
\section{Functions}{
\itemize{
\item \code{xml_attr_first}: Extract first xml attribute as vector
}}

\examples{
aq <- pubmed_nodeset(aqc)
xml_text_first(aq, "//Author/LastName")
}
